Fix type of integer field to bigint

+++ db/migrate/20201114125813_create_txs.rb
@@ -6,11 +6,11 @@ class CreateTxs < ActiveRecord::Migration[6.0]
       t.integer :size, null: false
       t.integer :vsize, null: false
       t.integer :weight, null: false
-      t.integer :locktime, null: false
+      t.bigint :locktime, null: false
       t.string :hex, null: false
       t.string :blockhash, null: false
-      t.integer :time, null: false
-      t.integer :blocktime, null: false
+      t.bigint :time, null: false
+      t.bigint :blocktime, null: false
       t.belongs_to :block
 
       t.timestamps
